% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_pair}
\alias{compute_pair}
\title{Single pair contribution}
\usage{
compute_pair(
  A,
  C_VEC,
  THETA,
  CORRFLAG,
  k,
  l,
  PAIRS_TABLE,
  SILENTFLAG,
  GRADFLAG
)
}
\description{
Wrapper of pair_contribution() used for unit tests
}
