% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_compute_pair_ext}
\alias{cpp_compute_pair_ext}
\title{Single pair contribution}
\usage{
cpp_compute_pair_ext(
  CONSTRMAT,
  CONSTRLOGSD,
  C_VEC,
  THETA,
  CORRFLAG,
  NTHR,
  NLOAD,
  NCORR,
  NVAR,
  K,
  L,
  PAIRS_TABLE,
  SILENTFLAG,
  GRADFLAG,
  OPTION = 0L
)
}
\arguments{
\item{C_VEC}{Vector containing the number of categories for each item}

\item{THETA}{Parameter vector}

\item{CORRFLAG}{1 to estimate latent correlations. 0 for orthogonal latent factors.}

\item{PAIRS_TABLE}{output from \code{\link[=pairs_freq]{pairs_freq()}}}

\item{SILENTFLAG}{optional for verbose output}

\item{GRADFLAG}{1 to compute gradient}

\item{A}{Constraint matrix. Loadings free to be estimated are identified by a 1.}

\item{k}{first index identifying the pair}

\item{l}{second index identifying the pair}
}
\description{
Wrapper of pair_contribution() used for unit tests
}
