% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_get_par_from_S}
\alias{cpp_get_par_from_S}
\title{Get transformed parameters from latent correlation matrix}
\usage{
cpp_get_par_from_S(S)
}
\arguments{
\item{S}{Latent correlation matrix.}
}
\description{
Use Lewandowski-Kurowicka-Joe (2009) transformation on latent correlations
with Choleski decomposition.
}
