% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_H}
\alias{estimate_H}
\title{Estimate of H}
\usage{
estimate_H(C_VEC, A, THETA, FREQ, N, CORRFLAG)
}
\description{
Compute a sample estimate of the expected negative Hessian by taking
advantage of the second Bartlett's identity at the single pair level
}
