% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_J}
\alias{estimate_J}
\title{Estimate of J}
\usage{
estimate_J(Y, C_VEC, A, THETA, CORRFLAG)
}
\description{
Compute a sample estimate of the variability matrix via the sample average outer product
of the composite score
}
