% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{extract_par}
\alias{extract_par}
\title{Extract parameters}
\usage{
extract_par(
  THETA,
  OPTION = c("list", "raw", "transformed"),
  CONSTRMAT,
  CONSTRLOGSD,
  LLC,
  NTHR,
  NLOAD,
  NCORR,
  NVAR
)
}
\arguments{
\item{THETA}{Unconstrained parameter vector.}

\item{OPTION}{Can take values \tabular{ll}{
\code{'raw'} \tab Returns \code{THETA}. \cr
\tab \cr
\code{'transformed'} \tab Transformed vector of parameters, with interpretable latent correlations. \cr
\tab \cr
\code{'list'} \tab List containing the ordered thresholds vector, the loading matrix and the latent correlation matrix. \cr
}}

\item{CONSTRMAT}{\eqn{p*q}-dimensional matrix. Elements set to \code{NA} refers to free loading parameters. Elements set to numerical values denote fixed values constraints.}

\item{CONSTRLOGSD}{\eqn{q}-dimensional vector. Elements set to \code{NA} refers to free latent log standard deviations parameters. Elements set to numerical values denote fixed values constraints.}

\item{LLC}{Linear loadings constraints. Expects a list of constraints. See \link{fit_plFA} documentation.}

\item{NTHR}{Number of thresholds parameters.}

\item{NLOAD}{Number of free loadings parameters}

\item{NCORR}{Number of free latent correlations parameters.}

\item{NVAR}{Number of free latent variance parameters.}
}
\description{
Extract parameters
}
