% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{extract_theta_path}
\alias{extract_theta_path}
\title{Extract parameters trajectory along the optimisation from
object of class StoFit}
\usage{
extract_theta_path(OBJ, LAB = "pathAvTheta")
}
\arguments{
\item{OBJ}{Object of class StoFit}

\item{LAB}{Can take values \code{pathTheta} for raw trajectories or \code{pathAvTheta} for averaged ones. Set by default at \code{pathAvTheta}.}
}
\description{
Extract parameters trajectory along the optimisation from
object of class StoFit
}
