% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plFA.R
\name{fit_plFA}
\alias{fit_plFA}
\title{Fit factor models for ordinal data with pairwise likelihood methods}
\usage{
fit_plFA(
  DATA_LIST = list("DATA", "CONSTRMAT", "CORRFLAG"),
  METHOD = "ucminf",
  CPP_CONTROL = list(),
  UCMINF_CONTROL = list(ctrl = list(), hessian = 0),
  INIT = NULL,
  ITERATIONS_SUBSET = NULL,
  VERBOSEFLAG = 0,
  NCORES = 1
)
}
\description{
fit_plFA() allows to fit models both numerically and stochastically
}
