% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lambda}
\alias{get_lambda}
\title{Extract free loadings from parameter vector}
\usage{
get_lambda(THETA, NTHR, NLOAD)
}
\arguments{
\item{THETA}{Unconstrained parameter vector}

\item{NTHR}{Number of thresholds parameters.}

\item{NLOAD}{Number of free loadings parameters}
}
\description{
Extract free loadings from parameter vector
}
