% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_S}
\alias{get_S}
\title{Extract latent covariance matrix from unconstrained parameter vector}
\usage{
get_S(THETA, CONSTRLOGSD, NTHR, NLOAD, NCORR, NVAR, Q)
}
\arguments{
\item{THETA}{Unconstrained parameter vector}

\item{CONSTRLOGSD}{\eqn{Q}-dimensional vector. Elements set to \code{NA} refers to free latent log standard deviations parameters. Elements set to numerical values denote fixed values constraints.}

\item{NTHR}{Number of thresholds parameters.}

\item{NLOAD}{Number of free loadings parameters}

\item{NCORR}{Number of free latent correlations parameters.}

\item{NVAR}{Number of free latent variance parameters.}

\item{Q}{Number of latent variables}
}
\value{
Latent covariance matrix of dimension  \eqn{Q*Q}.
}
\description{
Extract latent covariance matrix from unconstrained parameter vector
}
