% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tidy_path}
\alias{get_tidy_path}
\title{Extract estimation trajectories}
\usage{
get_tidy_path(MOD_OBJ, PATH_LAB)
}
\arguments{
\item{MOD_OBJ}{Output of \link{fit_plFA}.}

\item{PATH_LAB}{Label for the trajectory of interest. Use "path_av_theta"
for the averaged parameters trajectories.}
}
\description{
Extract trajectories of interest along the stochastic optimisation
from a fitted object
}
