% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{getPar,PlFaFit-method}
\alias{getPar,PlFaFit-method}
\title{Extract parameters}
\usage{
\S4method{getPar}{PlFaFit}(OBJ, ...)
}
\arguments{
\item{OBJ}{Object of class PlFaFit.}

\item{...}{Additional arguments}
}
\description{
Extract parameters
}
