% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{getPar,vector-method}
\alias{getPar,vector-method}
\title{Extract parameters}
\usage{
\S4method{getPar}{vector}(OBJ, ...)
}
\arguments{
\item{OBJ}{Raw theta vector to extract \code{'transformed'} and \code{'list'}.}

\item{...}{Additional arguments}
}
\description{
Extract parameters
}
