% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{getPar}
\alias{getPar}
\title{Extract parameters}
\usage{
getPar(OBJ, OPTION, ...)
}
\arguments{
\item{OBJ}{Object of class StoFit, PlFaFit. Also usable on raw theta vector to extract \code{'transformed'} and \code{'list'}.}

\item{OPTION}{Can take values \tabular{ll}{
\code{'raw'} \tab Raw vector of parameters. Latent correlations are reparametrised, thus not directly interpretable. \cr
\tab \cr
\code{'transformed'} \tab Transformed vector of parameters, with interpretable latent correlations. \cr
\tab \cr
\code{'list'} \tab List containing the ordered thresholds vector, the loading matrix and the latent correlation matrix. \cr
}}
}
\description{
Extract parameters
}
