% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{getThetaPath,PlFaFit-method}
\alias{getThetaPath,PlFaFit-method}
\title{Extract parameters trajectory along the optimisation}
\usage{
\S4method{getThetaPath}{PlFaFit}(OBJ, LAB = "pathAvTheta", OPTION = "transformed")
}
\arguments{
\item{OBJ}{Object of class StoFit}

\item{LAB}{Can take values \code{pathTheta} for raw trajectories or \code{pathAvTheta} for averaged ones. Set by default at \code{pathAvTheta}.}

\item{OPTION}{Parameterisation option.}
}
\description{
Extract parameters trajectory along the optimisation
}
