% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_S2}
\alias{grad_S2}
\title{Get latent correlation matrix from theta}
\usage{
grad_S2(A, THETA, IDX)
}
\arguments{
\item{THETA}{Numerical vector of parameters.}

\item{Q}{Number of latent variables.}
}
\description{
get_Sigma_u() extracts the latent correlation matrix from theta assuming
theta elements to be reparametrised following the
Lewandowski-Kurowicka-Joe (2009) transform.
}
