% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAT.R
\docType{data}
\name{LSAT}
\alias{LSAT}
\title{The Law School Admission Test (LSAT), Section VI}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 5 columns.
}
\source{
This LSAT example is a part of a data set given in Bock and Lieberman (1970).
}
\usage{
LSAT
}
\description{
The LSAT is a classical example in educational testing for measuring
  ability traits. This test was designed to measure a \emph{single}
  latent ability scale.
}
\references{
Bartholomew, D., Steel, F., Moustaki, I. and Galbraith, J. (2002)
\emph{The Analysis and Interpretation of Multivariate Data for
Social Scientists}. London: Chapman and Hall.

Bock, R. and Lieberman, M. (1970) Fitting a response model for \eqn{n}
dichotomously scored items. \emph{Psychometrika}, \bold{35}, 179--197.
}
\keyword{datasets}
