% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiThread_completePairwise}
\alias{multiThread_completePairwise}
\title{Complete pairiwse iteration with multithreading option//'
Used by external optimisers}
\usage{
multiThread_completePairwise(
  N,
  C_VEC,
  CONSTRMAT,
  THETA,
  FREQ,
  CORRFLAG,
  GRFLAG = 1L,
  SILENTFLAG = 1L
)
}
\description{
Complete pairiwse iteration with multithreading option//'
Used by external optimisers
}
