% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiThread_completePairwise_nll}
\alias{multiThread_completePairwise_nll}
\title{Full pairwise likelihood}
\usage{
multiThread_completePairwise_nll(
  N,
  C_VEC,
  CONSTRMAT,
  THETA,
  FREQ,
  CORRFLAG,
  SILENTFLAG
)
}
\arguments{
\item{N}{Number of observations}

\item{C_VEC}{Vector containing the number of categories for each item}

\item{CONSTRMAT}{Constraint matrix. Loadings free to be estimated are identified by a 1.}

\item{THETA}{Parameter vector}

\item{FREQ}{Frequency table}

\item{CORRFLAG}{1 to estimate latent correlations. 0 for orthogonal latent factors.}

\item{SILENTFLAG}{optional for verbose output}
}
\description{
Evaluate negative loglikelihood. Same structure of \code{multiThread_completePairwise}.
Used to monitor validation log-likelihood. Multithreading options via \code{RcppParallel}.
}
