% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairs_freq}
\alias{pairs_freq}
\title{Compute pairwise frequencies}
\usage{
pairs_freq(Y, C_VEC)
}
\arguments{
\item{Y}{Integer matrix of dimension \eqn{n*p}, where \eqn{n} is the sample size
and \eqn{p} is the number of items considered. Categories must be coded starting from zero.
For example, an item with three categories can only accept values contained in
\eqn{\{0, 1, 2\}}.}

\item{C_VEC}{Integer vector indicating how many possible categories are associated to
each item in 'Y'.}
}
\value{
It returns a 5-rows matrix with each combination of items and categories as columns.
Row0: item k, Row1: item l, Row2; category item k, Row3: category item l, Row4: freq
It is computed just once, before the optimization of the complete pairwise
}
\description{
Compute pairwise frequencies
}
