% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plFA}
\alias{plFA}
\title{Stochastic optimiser}
\usage{
plFA(
  FREQ,
  VALFREQ,
  N,
  C_VEC,
  CONSTRMAT,
  THETA_INIT,
  CORRFLAG,
  METHODFLAG,
  PAIRS_PER_ITERATION,
  ETA,
  BURN,
  MAXT,
  TOLCOUNT = 50L,
  SILENTFLAG = 1L,
  TOL = 1e-06,
  EACHCLOCK = 100L,
  PAR1 = 1,
  PAR2 = 1,
  PAR3 = 0.75,
  STEPSIZEFLAG = 0L,
  CHECKCONV = 0L,
  EACHCHECK = 100L,
  SEED = 123L
)
}
\arguments{
\item{FREQ}{Frequency table.}

\item{VALFREQ}{Frequency table validation data.}

\item{N}{Number of observations.}

\item{C_VEC}{Vector containing the number of categories for each item.}

\item{CONSTRMAT}{Constraint matrix. Loadings free to be estimated are identified by a 1.}

\item{THETA_INIT}{Initial parameter vector}

\item{CORRFLAG}{1 to estimate latent correlations. 0 for orthogonal latent factors.}

\item{METHODFLAG}{0 for hypergeometric, 1 for Bernoulli.}

\item{PAIRS_PER_ITERATION}{Number of pairs to draw per iteration.}

\item{ETA}{Initial stepsize.}

\item{BURN}{Initial burn-in period.}

\item{MAXT}{Maximum number of iterations.}

\item{TOLCOUNT}{Tolerance count for convergence with validation nll.}

\item{SILENTFLAG}{Silent output.}

\item{TOL}{Tolerance level.}

\item{EACHCLOCK}{How often (in terms of iterations) to measure single iteration computational times (using \code{RcppClock}..}

\item{PAR1}{Hyperparameter for stepsize scheduling by Xu (2011): Scaling.}

\item{PAR2}{Hyperparameter for stepsize scheduling by Xu (2011): Smallest Hessian eigenvalue.}

\item{PAR3}{Hyperparameter for stepsize scheduling by Xu (2011): Decay rate.}

\item{STEPSIZEFLAG}{Choose stepsize scheduling: Set 0 for Polyak and Juditsky (1992), 1 for Xu (2011).}

\item{CHECKCONV}{Flag to check for convergence using complete pairwise likelihood on the validation set.}

\item{EACHCHECK}{How often (in terms of iterations) to check for convergence.}

\item{SEED}{Randomising seed.}
}
\description{
Core function to evaluate stochastic estimators
}
