% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{show,PlFaFit-method}
\alias{show,PlFaFit-method}
\title{Show method for the output of fit_plFA()}
\usage{
\S4method{show}{PlFaFit}(object)
}
\arguments{
\item{object}{Object of class StoFit}
}
\description{
Show method for the output of fit_plFA()
}
