% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{sim_data}
\alias{sim_data}
\title{Simulate new data}
\usage{
sim_data(SAMPLE_SIZE, LOADINGS, THRESHOLDS, LATENT_COV, SEED = 123)
}
\arguments{
\item{SAMPLE_SIZE}{Number of observations to simulate.}

\item{LOADINGS}{Matrix of loadings, with dimensions \eqn{p*q}.}

\item{THRESHOLDS}{Vector of thresholds fo a single item.
It is assumed to be the same for all the items.}

\item{LATENT_COV}{Latent correlation matrix. Dimension \eqn{q*q}.}

\item{SEED}{Numeric seed for random generation.}
}
\description{
Simulate data given the values of the parameters for thresholds, loadings and latent correlations
}
