% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plFA.R
\name{stoc_args}
\alias{stoc_args}
\title{Internal function to check arguments in fit_plFA}
\usage{
stoc_args(ARGS, P)
}
\arguments{
\item{ARGS}{List of arguments}

\item{P}{Number of pairs}
}
\description{
Internal function to check arguments in fit_plFA
}
