% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{build_constrMat}
\alias{build_constrMat}
\title{Automatic construction of simple constraints matrices}
\usage{
build_constrMat(
  P,
  Q,
  STRUCT = c("simple", "triangular", "crossed"),
  CROSS = NULL
)
}
\arguments{
\item{P}{Number of items.}

\item{Q}{Number of latent vaiables.}

\item{STRUCT}{Label for the constraint structure chosen. Only two possible values:
'triangular', 'simple' and 'crossed'.}

\item{CROSS}{Integer referring to the number of items latent i shares with latent i+1}
}
\description{
build_constrMat() creates a binary matrix with entries equal to 1
when referring to free loadings, and 0 entries otherwise.
}
