% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{cfa}
\alias{cfa}
\title{Fit Confirmatory Factor Analysis Models using plFA}
\usage{
cfa(
  model,
  data,
  std.lv = TRUE,
  start = NULL,
  estimator = "PML",
  estimator.args = list(method = "ucminf", valdata = NULL, iterations_subset = NULL,
    ncores = 1),
  information = "observed",
  control = list(),
  ...
)
}
\arguments{
\item{model}{A description of the user-specified model. Typically, the model
is described using the lavaan model syntax. See \code{\link[lavaan:model.syntax]{lavaan::model.syntax}}
for more information. Alternatively, a parameter table (eg. the output of
the \code{\link[lavaan:model.syntax]{lavaan::lavaanify()}} function) is also accepted.}

\item{data}{A data frame containing the observed variables used in the model.
Variables must be declared as ordered factors.}

\item{start}{Starting values to use.}

\item{estimator}{(For \code{lavaan} compatibility only). The estimator is \code{PML}.}

\item{estimator.args}{A list of arguments for \code{\link[=fit_plFA]{fit_plFA()}}. TBC.}

\item{information}{(For \code{lavaan} compatibility only). The information matrix is...}

\item{control}{A list of control parameters for the estimation algorithm. See \code{\link[=fit_plFA]{fit_plFA()}} for more information.}

\item{...}{Additional arguments to be passed to \code{\link[=lavaan]{lavaan()}}.}
}
\value{
A \code{plFAlavaan} object.
}
\description{
Fit Confirmatory Factor Analysis Models using plFA
}
