% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_multiThread_completePairwise}
\alias{cpp_multiThread_completePairwise}
\title{Full pairwise iteration}
\usage{
cpp_multiThread_completePairwise(
  N,
  C_VEC,
  CONSTRMAT,
  CONSTRLOGSD,
  THETA,
  FREQ,
  CORRFLAG,
  NTHR,
  NLOAD,
  NCORR,
  NVAR,
  GRFLAG = 0L,
  SILENTFLAG = 1L
)
}
\arguments{
\item{N}{Number of observations}

\item{C_VEC}{Vector containing the number of categories for each item}

\item{CONSTRMAT}{\eqn{p*q}-dimensional matrix. Elements set to \code{NA} refers to free loading parameters. Elements set to numerical values denote fixed values constraints.}

\item{CONSTRLOGSD}{\eqn{q}-dimensional vector. Elements set to \code{NA} refers to free latent log standard deviations parameters. Elements set to numerical values denote fixed values constraints.}

\item{THETA}{Parameter vector}

\item{FREQ}{Frequency table}

\item{CORRFLAG}{1 to estimate latent correlations. 0 for orthogonal latent factors.}

\item{NTHR}{Number of thresholds parameters.}

\item{NLOAD}{Number of free loadings parameters}

\item{NCORR}{Number of free latent correlations parameters.}

\item{NVAR}{Number of free latent variance parameters.}

\item{GRFLAG}{0 to only compute the likelihood. 1 to also compute the gradient.}

\item{SILENTFLAG}{optional for verbose output}
}
\description{
Evaluate negative log pairwise likelihood or gradient of the complete pool of pairs.
Used by external optimisers. Multithreading options via \code{RcppParallel}.
}
