% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_H}
\alias{estimate_H}
\title{Estimate of H}
\usage{
estimate_H(
  C_VEC,
  A,
  CONSTRLOGSD,
  THETA,
  FREQ,
  N,
  CORRFLAG,
  NTHR,
  NLOAD,
  NCORR,
  NVAR
)
}
\arguments{
\item{C_VEC}{Vector containing the number of categories for each item}

\item{A}{Constraint matrix. Loadings free to be estimated are identified by a 1.}

\item{CONSTRLOGSD}{\eqn{q}-dimensional vector. Elements set to \code{NA} refers to free latent log standard deviations parameters. Elements set to numerical values denote fixed values constraints.}

\item{THETA}{Parameter vector}

\item{FREQ}{output from \code{\link[=pairs_freq]{pairs_freq()}}}

\item{N}{sample size}

\item{CORRFLAG}{1 to estimate latent correlations. 0 for orthogonal latent factors.}

\item{NTHR}{Number of thresholds parameters.}

\item{NLOAD}{Number of free loadings parameters}

\item{NCORR}{Number of free latent correlations parameters.}

\item{NVAR}{Number of free latent variance parameters.}
}
\description{
Compute a sample estimate of the expected negative Hessian by taking
advantage of the second Bartlett's identity at the single pair level
}
