% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_J}
\alias{estimate_J}
\title{Estimate of J}
\usage{
estimate_J(Y, C_VEC, A, CONSTRLOGSD, THETA, CORRFLAG, NTHR, NLOAD, NCORR, NVAR)
}
\arguments{
\item{Y}{data matrix}

\item{C_VEC}{Vector containing the number of categories for each item}

\item{A}{Constraint matrix. Loadings free to be estimated are identified by a 1.}

\item{THETA}{Parameter vector}

\item{CORRFLAG}{1 to estimate latent correlations. 0 for orthogonal latent factors.}
}
\description{
Compute a sample estimate of the variability matrix via the sample average outer product
of the composite score
}
