% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_J}
\alias{estimate_J}
\title{Estimate of J}
\usage{
estimate_J(
  Y,
  C_VEC,
  A,
  CONSTRLOGSD,
  LLC,
  THETA,
  CORRFLAG,
  NTHR,
  NLOAD,
  NCORR,
  NVAR
)
}
\arguments{
\item{Y}{data matrix}

\item{C_VEC}{Vector containing the number of categories for each item}

\item{A}{Constraint matrix. Loadings free to be estimated are identified by a 1.}

\item{CONSTRLOGSD}{\eqn{q}-dimensional vector. Elements set to \code{NA} refers to free latent log standard deviations parameters. Elements set to numerical values denote fixed values constraints.}

\item{LLC}{Linear loadings constraints. Expects a list of constraints. See \link{fit_plFA} documentation.}

\item{THETA}{Parameter vector}

\item{CORRFLAG}{1 to estimate latent correlations. 0 for orthogonal latent factors.}

\item{NTHR}{Number of thresholds parameters.}

\item{NLOAD}{Number of free loadings parameters}

\item{NCORR}{Number of free latent correlations parameters.}

\item{NVAR}{Number of free latent variance parameters.}
}
\description{
Compute a sample estimate of the variability matrix via the sample average outer product
of the composite score
}
