% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\name{extract_par}
\alias{extract_par}
\title{Extract parameters}
\usage{
extract_par(THETA, OPTION, C, P, Q, CONSTRMAT)
}
\arguments{
\item{THETA}{Raw theta vector.}

\item{OPTION}{Can take values \tabular{ll}{
\code{'raw'} \tab Returns \code{THETA}. \cr
\tab \cr
\code{'transformed'} \tab Transformed vector of parameters, with interpretable latent correlations. \cr
\tab \cr
\code{'list'} \tab List containing the ordered thresholds vector, the loading matrix and the latent correlation matrix. \cr
}}

\item{C}{Sum of the number of categories for each item.}

\item{P}{Number if items}

\item{Q}{Number of latent variables}

\item{CONSTRMAT}{Matrix of dimension \eqn{p * q} with binary loading constraints. 1 for free loadings, 0 otherwise.}
}
\description{
Extract parameters
}
