% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plFA.R
\name{fit_plFA}
\alias{fit_plFA}
\title{Fit factor models for ordinal data with pairwise likelihood methods}
\usage{
fit_plFA(
  DATA,
  CONSTR_LIST,
  VALDATA = NULL,
  METHOD = c("ucminf", "SA"),
  CONTROL = list(),
  INIT = NULL,
  ITERATIONS_SUBSET = NULL,
  VERBOSE = FALSE,
  NCORES = 1
)
}
\arguments{
\item{DATA}{Integer data matrix of dimension \eqn{n*p}. Categories must be coded starting from zero.}

\item{CONSTR_LIST}{List of constraints. It must contain \tabular{ll}{
\code{CONSTRMAT} \tab \eqn{p*q} binary matrix. Elements set to 1 refers to free loding paameters. Elements set to 0 refer to null loadings. \cr
\tab \cr
\code{CORRFLAG} \tab Binary indicator. Set it to 0 if the latent variables are assumed to be independent. Set it 1 otherwise. \cr
}}

\item{VALDATA}{Data matrix used as validation set to check for convergence.}

\item{METHOD}{Label for the method chosen. Possible values are: \tabular{ll}{
\code{'ucminf'} \tab for estimation via the numerical optimiser from the \code{ucminf} package. \cr
\tab \cr
\code{'bernoulli'} \tab for stochastic estimation with weights following a Bernoulli distribution. \cr
\tab \cr
\code{'hyper'} \tab for stochastic estimation with weights following a multivariate hypergeometric distribution. \cr
}}

\item{CONTROL}{List of control options to pass to the optimiser: \tabular{ll}{
\code{control} \tab  See \code{ucminf} documentation \cr
\tab \cr
\code{hessian} \tab  See \code{ucminf} documentation \cr
\tab \cr
\code{PAIRS_PER_ITERATION} \tab  (Average) number of pairs to draw stochastically at each iteration with method (\code{'bernoulli'}) \code{'hyper'}.   \cr
\tab \cr
\code{MAXT} \tab  Maximum number of iterations.\cr
\tab \cr
\code{BURN} \tab  Length of the burning period.\cr
\tab \cr
\code{ETA} \tab  Initial stepsize.\cr
\tab \cr
\code{PAR1} \tab Hyperparameter for stepsize scheduling by Xu (2011): Scaling. \cr
\tab \cr
\code{PAR2} \tab Hyperparameter for stepsize scheduling by Xu (2011): Smallest Hessian eigenvalue. \cr
\tab \cr
\code{PAR1} \tab Hyperparameter for stepsize scheduling by Xu (2011): Decay rate. \cr
\tab \cr
\code{STEPSIZEFLAG} \tab Choose stepsize scheduling: Set 0 for Polyak and Juditsky (1992), 1 for Xu (2011). \cr
\tab \cr
\code{EACHCLOCK} \tab How often (in terms of iterations) to measure single iteration computational times (using \code{RcppClock}. \cr
\tab \cr
\code{CHECKCONV} \tab Flag to check for convergence using complete pairwise likelihood on the validation set. \cr
\tab \cr
\code{EACHCHECK} \tab How often (in terms of iterations) to check for convergence. \cr
\tab \cr
\code{TOL} \tab Tolerance between consecutive evaluations of the validation composite likelihood. \cr
\tab \cr
\code{TOLCOUNT} \tab How many checks the algorithm is allowed to accept if the validation negative composite likelihood starts increasing. \cr
\tab \cr
}}

\item{INIT}{Initialising vector. If not provided, optiisation starts at the zero vector.}

\item{ITERATIONS_SUBSET}{Vector with the indexes of iterations used to store trajectories.}

\item{NCORES}{Integer value setting the number of threads to carry out the estimation.}

\item{VERBOSEFLAG}{TRUE verbose output}
}
\description{
fit_plFA() allows to fit models both numerically and stochastically
}
