% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_corr}
\alias{get_corr}
\title{Extract vector of correlations from parameter vector}
\usage{
get_corr(THETA, Q, CORRFLAG)
}
\arguments{
\item{THETA}{Parameter vector}

\item{Q}{Number of latent variables.}
}
\description{
get_corr() extracts the latent correlations from the
parameter vector. The vector has length \eqn{q*(q-1)/2},
where \eqn{q} is the number of latent variables.
}
