% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lambda}
\alias{get_lambda}
\title{Extract free loadings from parameter vector}
\usage{
get_lambda(THETA, NTHR, NLOAD)
}
\arguments{
\item{THETA}{Raw vector of parameters.}

\item{C}{Sum of the number of categories for each item.}

\item{P}{Number if items}

\item{Q}{Number of latent variables}
}
\description{
get_lambda() extracts the vector of free loadings
from the parameter vector
}
