% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_theta}
\alias{get_theta}
\title{Get full parameter vector}
\usage{
get_theta(TAU, LOADINGS, LATENT_COV, CAT, A)
}
\arguments{
\item{TAU}{Numerical vector of thresholds parameters.
If all the \eqn{p} items have \eqn{c} possible categories, it contains
\eqn{p(c-1)} elements.}

\item{LOADINGS}{Numerical matrix of dimension \eqn{p*q},
where \eqn{q} is the number of latent variables.}

\item{LATENT_COV}{Correlation matrix of dimension \eqn{q}.}

\item{CAT}{Integer vector storing the possible number of categories
for each item. Values must be ordered following items order in the dataset.}

\item{A}{Binary matrix of dimension \eqn{p*q}. A cell equal to \eqn{1} indicates
that the corresponding element in the loading matrix is free to be estimated.
A cell equal to \eqn{0} fixes the corresponding element in the loading matrix
to \eqn{0}.}
}
\description{
Assemble the parameter vector starting from the threshold vector,
the latent covariance matrix and the loading matrix
}
