% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiThread_completePairwise}
\alias{multiThread_completePairwise}
\title{Complete pairiwse iteration with multithreading option//'
Used by external optimisers}
\usage{
multiThread_completePairwise(
  Y,
  C_VEC,
  A,
  TAU,
  LAMBDA,
  TRANSFORMED_RHOS,
  FREQ,
  CORRFLAG,
  GRFLAG,
  SILENTFLAG
)
}
\description{
Complete pairiwse iteration with multithreading option//'
Used by external optimisers
}
